Crate libradb

This crate provides LibraDB which represents physical storage of the core Libra data structures.

It relays read/write operations on the physical storage via schemadb to the underlying Key-Value storage system, and implements libra data structures on top of it.

Modules

backup

errors
This module defines error types used by LibraDB.

schema
This module defines representation of Libra core data structures at physical level via schemas that implement schemadb::schema::Schema.

test_helper
This module provides reusable helpers in tests.

Structs

LibraDB
This holds a handle to the underlying DB responsible for physical storage and provides APIs for access to the core Libra data structures.

Statics

LIBRA_STORAGE_CF_SIZE_BYTES

LIBRA_STORAGE_COMMITTED_TXNS

LIBRA_STORAGE_LATEST_TXN_VERSION

Functions

test_save_blocks_impl