Struct libradb::LibraDB

pub struct LibraDB { /* fields omitted */ }

Methods

impl LibraDB

pub fn new<P: AsRef<Path> + Clone>(db_root_path: P) -> Self

pub fn get_account_state_with_proof(
&self,
address: AccountAddress,
version: Version,
ledger_version: Version
) -> Result<(Option<AccountStateBlob>, AccountStateProof)>

Return the account state corresponding to the given version and account address with proof based on ledger_version

pub fn get_events_by_event_access_path(
_version: Version,
_access_path: AccessPath,
_start_event_seq_num: u64,
_ascending: bool,
_limit: usize,
_ledger_version: Version
) -> Result<Vec<(ContractEvent, EventProof)>>

Return events specified by access_path with sequence number in range designated by start_event_seq_num, ascending and limit. If ascending is true this query will return up to limit events that were emitted after start_event_seq_num. Otherwise it will return up to limit events in the reverse order. Both cases are inclusive.

pub fn get_txn_by_account_and_seq(
&self,
address: AccountAddress,
seq_num: u64,
ledger_version: Version
) -> Result<Option<(SignedTransaction, SignedTransactionProof)>>

Return a signed trasaction that is the seq_num-th transaction associated with the given account. If the signed transaction with given seq_num doesn't exist, return None.

pub fn get_latest_version(&self) -> Result<Version>

Get the latest version number available in the ledger.

pub fn get_txns_by_version(
&self,
start_version: Version,
limit: u64,
ledger_version: Version
) -> Result<Vec<(SignedTransaction, SignedTransactionProof)>>

Return up to limit signed transactions starting from start_version.

pub fn get_account_state_with_proof_by_state_root(
&self,
address: AccountAddress,
state_root: HashValue
) -> Result<(Option<AccountStateBlob>, SparseMerkleProof)>

pub fn get_executor_startup_info(&self) -> Result<Option<ExecutorStartupInfo>>

pub fn save_transactions(
&self,
txns_to_commit: &[TransactionToCommit],
ledger_info_with_sigs: &Option<LedgerInfoWithSignatures>
) -> Result<()>

Persist transactions. Called by Execution when either syncing nodes or committing blocks during normal operation. When ledger_info_with_sigs is provided, verify that the transaction accumulator root hash it carries is generated after the txns_to_commit are applied.

pub fn update_to_latest_ledger_and_get_proofs(
&self,
_client_known_version: u64,
request_items: Vec<RequestItem>
) -> Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, Vec<ValidatorChangeEventWithProof>)>

pub fn get_consensus_data(
&self
) -> Result<(Option<ConsensusState>, Vec<BlockData>, Vec<ConsensusQuorumCertificateData>)>

pub fn save_consensus_state(
&self,
consensus_state: ConsensusState
) -> Result<()>

pub fn save_consensus_block(&self, block_data: Vec<BlockData>) -> Result<()>

pub fn save_consensus_quorum_certificate(
&self,
qc: Vec<ConsensusQuorumCertificateData>
) -> Result<()>

pub fn delete_consensus_blocks_and_quorum_certificates(
&self,
block_ids: &[HashValue]
) -> Result<()>

pub fn get_transaction_list_with_proof(
&self,
start_version: Version,
limit: u64,
ledger_version: Version
) -> Result<TransactionListWithProof>