Struct libradb::LibraDB

pub struct LibraDB { /* fields omitted */ }

This holds a handle to the underlying DB responsible for physical storage and provides APIs for access to the core Libra data structures.

Methods

impl LibraDB

pub fn open<P: AsRef<Path> + Clone>(
db_root_path: P,
readonly: bool
) -> Result<Self>

pub fn new<P: AsRef<Path> + Clone>(db_root_path: P) -> Self

This creates an empty LibraDB instance on disk or opens one if it already exists.

pub fn get_epoch_change_ledger_infos(
&self,
start_epoch: u64,
end_epoch: u64
) -> Result<(Vec<LedgerInfoWithSignatures>, bool)>

Returns ledger infos reflecting epoch bumps starting with the given epoch. If there are no more than MAX_NUM_EPOCH_CHANGE_LEDGER_INFO results, this function returns all of them, otherwise the first MAX_NUM_EPOCH_CHANGE_LEDGER_INFO results are returned and a flag (when true) will be used to indicate the fact that there is more.

pub fn get_transaction_with_proof(
&self,
version: Version,
ledger_version: Version,
fetch_events: bool
) -> Result<TransactionWithProof>

pub fn update_to_latest_ledger(
&self,
client_known_version: Version,
request_items: Vec<RequestItem>
) -> Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, EpochChangeProof, AccumulatorConsistencyProof)>

This backs the UpdateToLatestLedger public read API which returns the latest LedgerInfoWithSignatures together with items requested and proofs relative to the same ledger info.

pub fn get_backup_handler(&self) -> BackupHandler

Gets an instance of BackupHandler for data backup purpose.

pub fn restore_account_state(
&self,
iter: impl Iterator<Item = (Vec<(HashValue, AccountStateBlob)>, SparseMerkleRangeProof)>,
version: Version,
expected_root_hash: HashValue
) -> Result<()>